Garbage-Classification-System/requirements.txt
tensorflow>=2.0.0
numpy
matplotlib
Pillow
h5py
jupyterlab
Flask
